8.21修复bug

+++ Riot/Modules/Wallet/AssetsDetail/model/YXAssetsDetailListModel.h
@@ -33,7 +33,7 @@ typedef NS_ENUM(NSInteger , YXAssetsDetailRecordsStatue) {
 @property (nonatomic , copy) NSString              * type;
 @property (nonatomic , copy) NSString              * addr;
 @property (nonatomic , copy) NSString              * message;
-@property (nonatomic , assign) NSInteger              amount;
+@property (nonatomic , assign) CGFloat              amount;
 @property (nonatomic , assign) CGFloat              fees;
 @property (nonatomic , copy) NSString              * coinDate;
 @property (nonatomic , copy) NSString              * baseSybol;
